Revert "Merge branch 'main' into feature/firebase"
This reverts commit c1b69a636d514a061b084923e6869b5ca0033abd, reversing
changes made to 4425654e687f51232d2abe39f497574d4f2b6bc7.
