+++ src/JS/createGallery.js
@@ -10,8 +10,11 @@ const handleChange = e => (query = e.target.value.trim());
 
 const genresMovieIds = () => {
   getGenresMovies();
+  // console.log('alert: ', alert);
   const genreIdsJson = sessionStorage.getItem('genresIds');
+  // console.log(genreIdsJson);
   const genreIdsParse = JSON.parse(genreIdsJson);
+  // console.log(genreIdsParse);
   return genreIdsParse;
 };
 const genreMovie = genresMovieIds();
@@ -29,7 +32,6 @@ const createGenreArray = genre_Ids => {
 
 const handleClick = async e => {
   e.preventDefault();
-  let page;
   page = 1;
   clearGallery();
   alert.innerHTML = '';
@@ -45,15 +47,14 @@ export const createGallery = data => {
       } else {
         src = `https://image.tmdb.org/t/p/w500/${movie.poster_path}`;
       }
-      const stringVote = String(movie.vote_average).slice(0, 3);
-      
+
       return `
         <div class="film-card">
         <a href=https://image.tmdb.org/t/p/original/${movie.poster_path}>
         <img class="film-img" src="${src}" alt="${movie.title}" loading="lazy" /> 
         <div class="info">
         <p class="info-title">${movie.original_title}</p>
-        <p class="info-text">${genresTitle} | ${movie.release_date.substring(0,4)} <span class="info-rating">${stringVote}</span></p>
+        <p class="info-text">${genresTitle} | ${movie.release_date} ${movie.vote_average}</p>
         </div>
         </a>
         </div>
