+++ src/JS/renderingPopularMovies.js
@@ -18,11 +18,3 @@ getPopular().then(data => {
     });
   });
 });
-
-
-window.onload = function() {
-    if(!window.location.hash) {
-        window.location = window.location + '#loaded';
-        window.location.reload();
-    }
-}
