+++ src/library.html
@@ -3,9 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>F I L M O T E K A</title>
+    <title>Page title</title>
     <link rel="stylesheet" href="./sass/index.scss" />
-    <link rel="shortcut icon" href="./images/gitcat.gif" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -23,14 +22,13 @@
   </head>
   <body>
     <include src="./partials/library-header.html"></include>
-    <include src="./partials/gallery-library.html"></include>
     <include src="./partials/empty-library.html"></include>
     <include src="./partials/footer.html"></include>
     <include src="./partials/auth-menu.html"></include>
     <include src="./partials/signup-modal.html"></include>
     <include src="./partials/login-modal.html"></include>
     <div class="loader" id="loader"></div>
-    <script type="module" src="./library.js"></script>
+    <script type="module" src="./index.js"></script>
     <script
       src="https://kit.fontawesome.com/2707afadc2.js"
       crossorigin="anonymous"
