--- src/library.js
@@ -1,9 +1,0 @@
-import './JS/loader.js';
-import './JS/darkmode.js';
-
-// Firebase
-import './JS/auth-menu.js';
-import './JS/firebase-app.js';
-import './JS/signup.js';
-import './JS/login.js';
-// Firebase End
