+++ src/partials/empty-library.html
@@ -1,11 +1,8 @@
 <div class="empty__library">
-  <h2 class="empty__text1">Drogi kinomaniaku</h2>
-  <p class="empty__text2">Twoja biblioteka jest pusta dodaj jakieś filmy</p>
-  <div class="empty__john">
-    <img
-      class="empty__img"
-      src="https://media.tenor.com/FlUqvR77ru8AAAAd/bovisa-politecnico.gif"
-      alt="Johny empty library"
-    />
-  </div>
+<h2 class="empty__text1">Drogi kinomaniaku</h2>
+<p class="empty__text2">Twoja biblioteka jest pusta dodaj jakieś filmy</p>
+<div class="empty__john">
+    <img src="https://media.tenor.com/FlUqvR77ru8AAAAd/bovisa-politecnico.gif" alt="Johny empty library">
+</div>
+
 </div>
