--- src/partials/gallery-library.html
@@ -1,1 +1,0 @@
-<div class="films-cards-set"></div>
