+++ src/partials/header.html
@@ -22,7 +22,7 @@
             <a href="library.html" class="header-nav-items__link">My library</a>
           </li>
           <li>
-            <a href="#" class="header-nav-items__link auth-menu-open-btn">
+            <a href="" class="header-nav-items__link auth-menu-open-btn">
               <svg class="user-icon" width="14" height="14">
                 <use href="./images/icons.svg#icon-user"></use>
               </svg>
@@ -52,11 +52,13 @@
 
     <!-- Dark Mode -->
     <div class="toggle__wrapper">
-      <div class="fa-regular fa-sun fa-lg" style="color: #ff6b08"></div>
-      <label class="toggle__switch">
-        <input type="checkbox" id="toggle__checkbox" />
-        <span class="toggle__slider"></span>
-      </label>
+      <span id="toggle__icon">
+        <i class="fa-regular fa-sun fa-lg" style="color: #ff6b08"></i>
+        <label class="toggle__switch">
+          <input type="checkbox" id="toggle__checkbox" />
+          <span class="toggle__slider"></span>
+        </label>
+      </span>
     </div>
   </div>
 </section>
