+++ src/partials/library-header.html
@@ -20,7 +20,7 @@
             >
           </li>
           <li>
-            <a href="#" class="nav__link auth-menu-open-btn">
+            <a href="" class="nav__link auth-menu-open-btn">
               <svg class="user-icon" width="14" height="14">
                 <use href="./images/icons.svg#icon-user"></use>
               </svg>
@@ -45,11 +45,13 @@
     </div>
     <!-- Dark Mode -->
     <div class="toggle__wrapper">
-      <div class="fa-regular fa-sun fa-lg" style="color: #ff6b08"></div>
-      <label class="toggle__switch">
-        <input type="checkbox" id="toggle__checkbox" />
-        <span class="toggle__slider"></span>
-      </label>
+      <span id="toggle__icon">
+        <i class="fa-regular fa-sun fa-lg" style="color: #ff6b08"></i>
+        <label class="toggle__switch">
+          <input type="checkbox" id="toggle__checkbox" />
+          <span class="toggle__slider"></span>
+        </label>
+      </span>
     </div>
   </div>
 </section>
