+++ src/sass/_empty-library.scss
@@ -1,37 +1,42 @@
 .empty__library {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
 }
 
 .empty__text2 {
-  font-size: 16px;
-  color: lightcoral;
-  font-weight: bold;
+    font-size: 16px;
+    color: lightcoral;
+    font-weight: bold;
 }
 
 .empty__text1 {
-  padding: 10px;
-  font-size: 1.5rem;
+    padding: 10px;
+    font-size: 1.5rem;
 }
 
-.empty__img {
-  padding: 0 10px 10px 10px;
+img {
+    padding: 0 10px 10px 10px;
 }
-@media screen and (min-width: $tablet) {
-  .empty__text1 {
+@media  (min-width: $tablet){
+   
+
+
+.empty__text1 {
     padding: 10px;
     font-size: 2rem;
-  }
+}
 
-  .empty__text2 {
+.empty__text2 {
     font-size: 24px;
-  }
 
-  .empty__img {
+}
+
+img {
     padding-bottom: 10px;
     width: 100%;
     height: 100%;
-  }
+}
+
 }
