+++ src/sass/_footer.scss
@@ -25,9 +25,9 @@
   }
 }
 
-// .container.footer {
-//   background-color: $footer;
-// }
+.container.footer {
+  background-color: $footer;
+}
 
 .footer__wrapper {
   align-items: center;
