+++ src/sass/_gallery.scss
@@ -32,24 +32,11 @@
   color: $black;
 }
 .info-text {
-  display: flex;
-  align-items: center;
-  gap: 8px;
   font-style: normal;
   font-weight: 500;
   font-size: 20px;
   line-height: 1.2;
   color: $orange-font;
-}
-.info-rating {
-  background: $orange-font;
-  border-radius: 5px;
-  padding: 3px 11px;
-  color: $white;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 1.16;
 }
 
 //TABLET STYLES
