+++ src/sass/_variable.scss
@@ -8,7 +8,6 @@ $orange-button: #ff6b01;
 $background: #e5e5e5;
 $footer: #f7f7f7;
 $footer-font: #545454;
-$dark-mode-background: #1a1818;
 
 //Punkty przerwania
 $mobile: 320px;
